update to of 0.8

+++ example-motion/src/testApp.h
@@ -36,7 +36,7 @@ class testApp : public ofBaseApp{
 
         bool live;
 
-        int inputWidht, inputHeight;
+        float inputWidht, inputHeight;
 
 };
